requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.1.1
Django==4.2.19
django-stubs==5.1.3
django-stubs-ext==5.1.3
exceptiongroup==1.2.2
executing==2.2.0
ipython==8.18.1
jedi==0.19.2
matplotlib-inline==0.1.7
mypy==1.15.0
mypy-extensions==1.0.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
python-dotenv==1.0.1
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
wcwidth==0.2.13
